Refactor dbt adapter
