+++ piperider_cli/cli.py
@@ -7,7 +7,8 @@ import sentry_sdk
 from rich.console import Console
 from rich.syntax import Syntax
 
-from piperider_cli import dbt_adapter, __version__, event
+from piperider_cli import __version__, event
+from piperider_cli.adapter.dbt_adapter import DbtAdapter
 from piperider_cli.assertion_generator import AssertionGenerator
 from piperider_cli.event.track import TrackCommand
 from piperider_cli.guide import Guide
@@ -119,7 +120,7 @@ def init(**kwargs):
         if dbt_project_dir:
             dbt_project_path = os.path.join(dbt_project_dir, "dbt_project.yml")
         else:
-            dbt_project_path = dbt_adapter.search_dbt_project_path()
+            dbt_project_path = DbtAdapter.search_dbt_project_path()
 
     if dbt_project_path:
         console.print(f'Use dbt project file: {dbt_project_path}')
